refactor: Rename 'User' to 'MyPage' for my page related file & package (#42)
- Rename `UserViewModel` to `MyPageViewModel` and update its package to `feature.mypage`.
- Rename package `feature.user` to `feature.mypage`.
- Move `ProductListItem` to `common.component` and rename to `CommonProductListItem`.
- Remove unused dependency and function in `MyPageViewModel`
